requirements.txt
flask==2.3.3
flask-cors==4.0.0
gunicorn==21.2.0
numpy==1.24.3
scikit-learn==1.3.2
xgboost==1.7.6
hmmlearn==0.3.2
joblib==1.3.2
lightgbm==3.3.5
requests==2.31.0
opencv-python==4.8.1.78
pillow==10.0.1
selenium==4.15.0
webdriver-manager==4.0.1
line-bot-sdk==2.4.1
python-dotenv==1.0.0